Always allow pasting reference
